add new filter: friendly_hostname
Add a filter which takes a string and generates a deterministic
"friendly hostname" from it in a way which has a low chance of
collision. The possible number of hostnames is equivalent to around 24
bits of entropy. We're probably OK for a few hundred hosts before the
Birthday paradox says we have a 1% chance of collision.