add udev configuration for "friendly" device names
Add configuration for udev which maps USB capture devices to the
friendly names /dev/video-{presenter,presentation} and update the
galicaster profiles to use these friendly names.

The udev configuration uses the physical USB path to determined which
device gets which name which means that individual USB ports become the
"presenter" input or "presentation" input. Which ports these are is
motherboard specific and is configured in the motherboard-specific
configuration.

The udev rules also include configuration for the separate audio devices
so that the devices listed by "arecord -l" have friendly names as well.